save y list ok

+++ src/main/java/com/gestion/hoteles/persistencia/UsuarioRepositorio.java
@@ -7,8 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface UsuarioRepositorio extends JpaRepository<Usuario,Long>{
-
+public interface UsuarioRepositorio extends JpaRepository<Usuario,Integer>{
 
 
 }
